Minor changes
Add UnitTests for Iterable examples
