+++ core-java-modules/core-java-collections-2/src/main/java/com/baeldung/collections/iterable/IterableExample.java
@@ -1,48 +1,32 @@
 package com.baeldung.collections.iterable;
 
-import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-class IterableExample {
+public class IterableExample {
 
-    public static void iterateUsingIterator() {
-        List<Integer> numbers = getNumbers();
-
+    public void iterateUsingIterator(List<Integer> numbers) {
         Iterator<Integer> iterator = numbers.iterator();
         while (iterator.hasNext()) {
             System.out.println(iterator.next());
         }
     }
 
-    public static void removeElementsUsingIterator() {
-        List<Integer> numbers = getNumbers();
-
-        Iterator<Integer> iterator = numbers.iterator();
-        while (iterator.hasNext()) {
-            iterator.remove();
-        }
-    }
-
-    public static void iterateUsingEnhancedForLoop() {
-        List<Integer> numbers = getNumbers();
-
+    public void iterateUsingEnhancedForLoop(List<Integer> numbers) {
         for (Integer number : numbers) {
             System.out.println(number);
         }
     }
 
-    public static void iterateUsingForEachLoop() {
-        List<Integer> numbers = getNumbers();
+    public void iterateUsingForEachLoop(List<Integer> numbers) {
         numbers.forEach(System.out::println);
     }
 
-    private static List<Integer> getNumbers() {
-        List<Integer> numbers = new ArrayList<>();
-        numbers.add(10);
-        numbers.add(20);
-        numbers.add(30);
-        numbers.add(40);
-        return numbers;
+    public void removeElementsUsingIterator(List<Integer> numbers) {
+        Iterator<Integer> iterator = numbers.iterator();
+        while (iterator.hasNext()) {
+            iterator.next();
+            iterator.remove();
+        }
     }
 }
